ScopeContext - Extracted ScopeContextAsyncState into NLog.Internal
